+++ src/main/java/net/staticstudios/prisons/gambling/Flip.java
@@ -3,13 +3,10 @@ package net.staticstudios.prisons.gambling;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 
-import java.math.BigInteger;
-
-
 public class Flip {
     public Player owner;
     public String uuid;
-    public BigInteger amount;
+    public long amount;
     public boolean isHeads;
     public Material headsIcon;
     public Material tailsIcon;
