+++ src/main/java/net/staticstudios/prisons/pickaxe/abilities/MeteorStrikeAbility.java
@@ -3,12 +3,10 @@ package net.staticstudios.prisons.pickaxe.abilities;
 import net.staticstudios.prisons.pickaxe.abilities.handler.BaseAbility;
 import net.staticstudios.prisons.utils.PrisonUtils;
 
-import java.math.BigInteger;
-
 public class MeteorStrikeAbility extends BaseAbility { //todo
 
     public MeteorStrikeAbility() {
-        super("meteorStrike", "&6&lMeteor Strike", 15, BigInteger.valueOf(15), 1000 * 60 * 300,
+        super("meteorStrike", "&6&lMeteor Strike", 15, 15, 1000 * 60 * 300,
                 "&oSend meteors through the mine",
                 "&odestroying all blocks in their path!",
                 "",
