+++ src/main/java/net/staticstudios/prisons/pickaxe/enchants/DoubleFortuneEnchant.java
@@ -1,15 +1,13 @@
 package net.staticstudios.prisons.pickaxe.enchants;
 
-import net.kyori.adventure.text.format.NamedTextColor;
 import net.staticstudios.prisons.blockBroken.BlockBreak;
 import net.staticstudios.prisons.pickaxe.enchants.handler.BaseEnchant;
 import net.staticstudios.prisons.pickaxe.enchants.handler.EnchantTier;
 
-import java.math.BigInteger;
 
 public class DoubleFortuneEnchant extends BaseEnchant {
     public DoubleFortuneEnchant() {
-        super("doubleFortune", "&d&lOre Splitter", 1000, BigInteger.valueOf(5000), "&7Chance to get 2x blocks from fortune");
+        super("doubleFortune", "&d&lOre Splitter", 1000, 5000, "&7Chance to get 2x blocks from fortune");
 
         setTiers(
                 new EnchantTier(5, 0),
